+++ src/gort/gort.py
@@ -497,7 +497,6 @@ class Gort(GortClient):
                 park_telescopes=True,
                 disable_overwatcher=True,
                 show_message=False,
-                reset_lockout=True,
             )
 
         except Exception:
@@ -509,7 +508,6 @@ class Gort(GortClient):
                 park_telescopes=False,
                 force=True,
                 mode="overcurrent",
-                reset_lockout=True,
             )
 
     async def observe(