+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Removed the `reset_lockout` argument for the shutdown recipe since it's not required to allow the dome to move.
 * Do not close the dome during daytime if the Overwatcher is disabled.
+* Remove dome lockout after unsafe conditions. It may be reimplemented differently in the future.
 
 
 ## 1.6.0 - January 26, 2025